fixed bug #2010261: Java.flex example file can't handle most negative int literal
Integer literals (dec, oct, hex) can now start with an optional "-". This only works if there is no whitespace between the minus and the digits. My reading of the standard is that -  2147483648 (with a white space between the minus and the number) is to be lexed as unary minus and a (too large) int literal and should therefore be rejected.
 

git-svn-id: https://jflex.svn.sourceforge.net/svnroot/jflex/trunk@381 670313ca-f036-0410-890c-e3abb03e7d8f